conductor: use port binding extended API in during live migrate
This adds the changes to the conductor live migrate task such
that if the following conditions are met:

* Using neutron
* Port binding extended extension is available
* Both the source and dest host are new enough

Then port bindings are created in neutron for the destination
host and the information is stored in the migrate_data object
which is passed around the computes during live migration.

This will "turn on" the code in the compute service and virt
driver to do things like change vif types between hosts during
a live migration and reduce the network downtime after the
guest is migrated since the port is already bound on the
destination host before we even started migrating the guest.

Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>

Part of blueprint neutron-new-port-binding-api

Change-Id: Ie04ded7b0f4edbb98af92f2ce713cf712a83b383