Added mount fstype based validation of Quobyte mounts
The validation of Quobyte mounts is extended to validate based
on a mounts file system type being set to "fuse.quobyte".
This includes adding a new StaleVolumeMount exception type to
the Nova exceptions.
This also closes a bug concerning multi-registry configurations
for Quobyte volumes due to no longer using the is_mounted()
method that failed in that case.
Finally this adds exception handling for the unmount call that
is issued on trying to mount an already mounted volume.

Closes-Bug: #1730933
Closes-Bug: #1737131

Change-Id: Ia5a23ce1123a68608ee2ec6f2ac5dca02da67c59