Simplify BDM boot index checking
The previous code had a complex check which asserted that:

1. the boot index list contains 0
2. every element in the boot index list is 1 less than its successor

The second check was in a method called _subsequent_list, which I
assume was intended to be _sequential_list.

However, combined with the constraint that the boot index are >= 0,
the above 2 constraints are the same as asserting that each element in
the boot index list has the same value as its index in the list. This
is also much simpler to understand.

This was originally submitted as change I8a3e7e6c, but was reverted in
change I299bf846 as it failed to consider that there may be no indices
at all. The revert added a test for this case.

Change-Id: I4fe228898f49e60a85a804d027a30d4c18b96887