Fix rebuild of baremetal instance when vm_state is ERROR
Nova allows rebuild of instance when vm_state is ERROR. [1]

The vm_state is restored to ACTIVE only after a successful build.
This means rebuilding a baremetal instance using the Ironic driver
is impossible because wait_for_active fails if vm_state=ERROR is found.

This is a regression introduced in a previous change which added
the ability to delete an instance in spawning state. [2]

This present change will skip the abort installation logic
if task_state is REBUILD_SPAWNING while preserving the previous logic.

[1] https://bugs.launchpad.net/nova/+bug/1183946
[2] https://bugs.launchpad.net/nova/+bug/1455000

Change-Id: I857ad7264f1a7ef1263d8a9d4eca491d6c8dce0f
Closes-bug: #1735009