Remove vestigial extra_info update in PciDevice.save()
The extra_info online migration was added in change
50355c4595e08f293f610da32247e405b20c1c5b and that online
data migration was removed in change
15cda7141ba88d2170f438eae5f7fba60f27fb2e but there is
one part where we're always putting the extra_info from
the object into the updates dict that goes to the
db.pci_device_update, which means we always update a
PciDevice even if nothing changes.

This removes that unnecessary update.

Change-Id: I18049dc97b662102ac6a5424eed4934dac9a2ce8