ProviderTree.get_provider_uuids()
As we start using nested resource providers in the resource tracker, and
virt drivers, we're going to need to be able to e.g. diff provider trees
and create or delete entries in placement accordingly.  This change set
implements get_provider_uuids() on ProviderTree, which returns a set of
the UUIDs of all providers, optionally at and below the level of a
provider specified by name or UUID.  So you can say e.g.:

 uuids_to_delete = old.get_provider_uuids() - new.get_provider_uuids()

or

 numa1_uuids = ptree.get_provider_uuids(name_or_uuid=numa_cell_1_uuid)

Change-Id: I994442830588ee37eda409370a07152903b2e817
blueprint: nested-resource-providers