Fix guestfs.set_backend_settings call
This method takes a list of strings. It was never failing
because python-libguestfs takes the input and casts it to
a list, which turns "force_tcg" into:

  ['f', 'o', 'r', 'c', 'e', '_', 't', 'c', 'g']

Which is obviously not correct.

http://libguestfs.org/guestfs.3.html#guestfs_set_backend_settings

This fixes that issue and leaves a TODO about investigating
the usage of set_backend_setting for just setting the
single force_tcg value instead of overrwriting all backend
settings.

Credit to Clark Boylan for identifying the issue.

Change-Id: I36edfaf30fd0cc234f929fd6937d07b6a1b646ae
Partial-Bug: #1735823