Fix cellsv1 messaging test
This test was completely not doing the thing it claimed to, because it
was calling the fake_save() method with keyword args that it wasn't
expecting. If you make the test fail for another reason, you see it in the
debug output, but not if it passes. Since the messenger swallows the
exception, we never notice that we did the wrong thing.

Change-Id: Id001f849709656c0ab30e8a8e8b8b3cc1d95d1d2