Add a new check to volume attach
With the new Cinder volume attach API we can attach the same volume
to the same instance multiple times. This is useful for live migrate
however in other cases we would like to avoid this to happen.

This patch adds a new check in Nova and fail the volume attach request
in case this condition is met.

Change-Id: I049f00f993e45eeb090a1e1a5e5696cf2f103187