Add new style volume attachment support to block_device.py
This plumbs in the new-style volume attachment support to the
virt driver block device code which does the actual work of
attaching the volume via the driver and completing the attachment
with cinder (the thing that makes the volume status "in-use").

At this point, none of the new flow is exercised outside of tests
because we are not yet setting the attachment_id when attaching
volumes, that comes in a later change to the API.

It's worth noting that when nova creates a volume during boot
from volume for a source_type=blank/image/snapshot BDM, we won't
have an attachment_id for that volume so we'll attach using the
old flow. For now this is OK, but might be something we need to
revisit in the future when we eventually want to remove the old
flow.

The structure here is to separate the old/new attach flows into
separate methods, despite quite a bit of copy/paste code in both
places. A single method using conditionals in different places
is arguably uglier (as seen in the test code) and we could likely
consolidate some of the common code into helper methods later.

Note that test_refresh_connection_info_with_attachment_id is
removed since it's now redundant with test_refresh_connection.

Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>

Part of blueprint cinder-new-attach-apis

Change-Id: Ib1b6b223c9d04579828d47607006ecd98b472e5a