Avoid stashed connector lookup for new style detach
We can avoid all of the stashed connector hulabaloo
when detaching a volume if we're using a new style
attachment, because we don't need a host connector
in that case, stashed or not, we just delete the
attachment and we're done. This also helps to avoid
some confusion about whether or not we need to stash
a connector in the connection_info for the new style
attachment flow (we don't).

Part of blueprint cinder-new-attach-apis

Change-Id: I1404842f91279dd40ec9e03ccbbfe11cfc48520c