Remove confusing comment in compute_node_get API method
HostMappings in the API database are created using the
compute_nodes.host field, and are not 1:1 with compute nodes
because you can have a 1:M host:node mapping in a cell
(thanks ironic!).

So the comment about putting the compute node uuid in the
host mapping to make things easier is incorrect and confusing
because HostMappings are really representing compute service
hosts, not compute nodes.

Change-Id: I08b2a5b4155319f67ee1feb1490421b4a0f4d069