Don't launch guestfs in a thread pool if guestfs.debug is enabled
When guestfs.debug is enabled, we're handling callback events
from guestfs and logging them at debug level. When guestfs
is launched to inspect capabilities, that is currently done
in an eventlet thread pool. Because of the concurrent logging
along with the eventlet thread, we can hit an issue where eventlet
tries to switch threads and fails and then we hang the launch
call to guestfs, which hangs creating an instance.

This change simply avoids using a thread pool to launch guestfs
if guestfs.debug is True.

Change-Id: I0ffe93a031154b123c8beff96a695df5a280b935
Closes-Bug: #1737214
(cherry picked from commit 7c30da13842469e641d171b6008a57d5db8c5d5e)