conf: Do not inherit image signature props with snapshots
All image signature properties should not be inherited from the metadata
of the original image when creating a snapshot of an instance. Otherwise
Glance will attempt to verify the signature of the snapshot image and
fail as this has changed from that of the original.

Closes-bug: #1737513
Needed-By: https://review.openstack.org/551290
Change-Id: Ia3d80bf2f81c7317fec117aecbc3c560d51a7d4e
(cherry picked from commit 82c823bcec5850f6876308ebcdbcb37312bf763f)