Update and complete volume attachments during resize
With the new cinder volume attachment flow, during a resize the
source compute will create a new volume attachment for any BDMs
connected to the instance and delete the existing volume attachments
which represents the connection to the source host. The finish_resize
flow on the destination compute will then refresh the connection info
on the BDMs it's working with before passing them to the virt driver.

Since the source compute updated the BDM.attachment_id to point at
the new attachment which is meant for the destination host, the
refresh_connection_info call will get the connection_info for the new
reserved attachment, which is actually empty since it hasn't been
connected to the destination host yet. This results in wiping out the
BDM.connection_info which has the "driver_volume_type" value which is
something the virt driver on the destination compute needs to know
which volume backend driver to use to connect the volume to the
destination host.

This change updates the volume attachments on the destination host
before refreshing the connection_info in the BDMs and before we call
the driver to finish the resize, and then we also complete the
volume attachments once the driver has successfully finished the
resize. This is similar to what would happen via _prep_block_device
during the initial instance create, but since the code paths into the
driver are different we have to handle this explicitly in the compute
manager.

Similarly, we have to perform the same incantation when reverting a
resize and going back to the original source host.

Closes-Bug: #1737599

Change-Id: Ifc80d07d94311534fd9e7824ede9d09223a011c2