Wait for live_migration_rollback.end notification
The test_live_migration_abort step of the test_migration_actions
notification sample test could randomly fail on the number of
notification received. This happens becuase the test only waits for
the live_migration_abort notification but then expect that also
the live_migration_rollback notification is received. However the
rollback notification is emited after the abort which sometimes
makes the test fast enough to miss that.

The patch adds an explicit wait for the rollback notification to
stabilize the test

Change-Id: I70d8f661fa75e7688e7e3cd527c9d1ee7415c50b
Closes-Bug: #1737781