Remove unnecessary connector stash in attachment_update
We don't need to stash the connector in the volume attachment
connection_info for new style volume attachments becaues it's
not used. This was copied from the initialize_connection method
which used it during detach of a volume, but for new style
volume attachments, we simply delete the volume attachment and
we don't need the original host connector.

Similarly, attachment_create wasn't stashing the connector either
if we created an attachment with a connector.

The two places that would use the stashed connector for an old
style attachment are:

1. DriverVolumeBlockDevice.detach
2. _get_stashed_volume_connector in the compute.API

But as noted, neither of those flows need the connector for a
new style attachment.

Change-Id: Ie028c26ea0436bfa0b613c9cffc6b6352792fc0a