Add nova-status check for ironic flavor migration
In Pike we started requiring that ironic instances have their
embedded flavor migrated to track the ironic node custom
resource class. This can be done either via the normal running
of the nova-compute service and ironic driver or via the
'nova-manage db ironic_flavor_migration' command.

This change adds a nova-status check to see if there are any
unmigrated ironic instances across all non-cell0 cells, and
is based mostly on the same logic within the nova-manage command
except it's multi-cell aware and doesn't use the objects.

Conflicts:
      doc/source/cli/nova-status.rst

NOTE(mriedem): The conflict is because the Rocky section
in the man page does not exist in Queens. The note about
the new check is added to the Queens section and mentions
it was backported from Rocky.

Change-Id: Ifd22325e849db2353b1b1eedfe998e3d6a79591c
(cherry picked from commit 7eb670352126ed8d30426eb422fe5aed9cb21e7f)