Retry _trait_sync on deadlock
We're seeing DBDeadlock failures during scheduling in CI jobs
when syncing traits when getting allocation candidates.

We have a lock around this code but that's not going to carry across
multiple processes, so we need to be able to retry on deadlock if
one occurs.

Change-Id: I6cf1793c1cbed18d850ec7e32b5b195e78cb4e68
Closes-Bug: #1738083
(cherry picked from commit c66ae65775bb9d885fac059847063fee70617bc5)