Implement set_and_clear_allocations in report client
set_and_clear_allocations is added to the scheduler's report client to
provide an atomic method which can set and clear allocations for two
different consumer uuids by making a POST to /allocations. Because we
have committed to using microversion 1.14 we do not try falling back
to an earlier version as the necessary functionality is provided in
microversion 1.13.

Three FIXMEs left behind when dansmith was doing the migration.uuid
with allocations work earlier in blueprint migration-allocations have
been changed from using those two fallback methods to using the
new one.

Also there are some concerns about the return values from the
methods. Are we doing enough with them?

Are there other places where this change should happen that I missed?

Is the new method too specific? Should it be more generic about
assembling aribtrary allocations that it is given?

Change-Id: I88aa32a95ea64ab7a9214772fd6e2a80c185d0cd