Expose driver_block_device fields consistently
Currently only _proxy_as_attr fields are available as attributes, and
they are not available as dict values. Other values are available as
both. This is inconsistent. This change ensures all fields are exposed
as both dict values and attributes, so the caller doesn't have to
remember which type of attribute they're accessing.

Part of bp local-disk-serial-numbers

Change-Id: I35ebff9a06617aa58e612919eba81e8eea21f6a3