Fix duplicate allocation candidates
Some combinations of resource providers were duplicated in
allocation candidates.

This can happen when multiple sharing RPs are in the same aggregates,
because both sharing RPs visit each other during the algorithm to
see if it can be used as an allocation candidate's mate.

This patch adds to _alloc_candidates_with_shared() function a check
if we already have that combination in allocation candidates by
adding `alloc_prov_ids`, which is a list of sets of allocation
candidates's rp_ids.

Change-Id: Iefd3137ec68029d27081a743e52e327625585081
Closes-Bug:#1730730