Remove unused argument from LibvirtDriver._disconnect_volume
_disconnect_volume was being passed disk_dev, which was is name of the
disk as presented to the gust. However, _disconnect_volume is only
concerned with unmounting the volume from the host, so this isn't
relevant.

A couple of volume drivers were using it for logging. We remove these
because it wasn't done consistently, is better done by the caller, and
isn't required as the information is available from other log
messages.

In the very minor associated debug log cleanup we also remove logging
of connection_info from one volume driver, which is potential security
bug.

We also do some associated cleanup in a few volume driver tests which
were assuming that disconnect_volume was being passed disk_info rather
than disk_dev, and that connection_info['data'] is related to
disk_info, which it isn't.

Change-Id: I61a0bee9e71e9a67f6a7c04a7bfd6e77fe818a77