api-ref: sort parameters for limits, quotas and quota classes
The limits, quotas and quota classes parameters tables are
getting a bit confusing now after several deprecation microversions.

This change sorts the parameters for those APIs such that the
deprecated parameters are pushed to the end of the parameter table,
sorted by microversion in ascending order and alphabetical within
the same microversion.

Change-Id: Icedabe222ed7e1ce9a5665e3aaeefb520b16cf83