Remove microversion fallback code from report client
For Queens, we've committed nova to using microversion 1.14 of the
placement API. That means the nova report client does not need to
fall back to earlier versions for some methods when removing
inventory, putting allocations, or ensuring that resource classes
exist.

This change removes fallback code that was only run when a 406
response was received from the placement service. Associated
tests that prove the falling back have been removed, and tests
that mocked methods that may have been fallen back have had
those mocks removed.

Change-Id: I2563213dc3b184049af336f0807c4cb4690a9f9a