Don't try to delete build request during a reschedule
If populate_retry failed because of MaxRetriesExceeded,
don't try to delete build requests because they should
already be gone from the initial create attempt, plus
we should assume the cell conductor can't reach the API
database anyway.

Similar for hitting NoValidHost during a reschedule. We
can tell if we're doing a reschedule by the num_attempts
value in filter_properties, populated via populate_retry,
which will be >1 during a reschedule.

Change-Id: I0b3ec6bb098ca32ffd32a61d4f9dcf426c3faf46
Closes-Bug: #1736946