Workaround missing RequestSpec.project_id when moving an instance
The online data migration routine to create request specs for old
instances used an admin context which has an empty project_id,
so when scheduling (moving) one of these, if we try to PUT /allocations
in placement using the FilterScheduler we'll fail because the project_id
is None.

This works around that by putting the instance.project_id on the request
spec before calling the scheduler to pick a node and claim resources
against it.

A later change will need to add some sort of online data migration
routine so that we properly update and persist the fix for these
older records.

Change-Id: I34b1d99a9d0d2aca80f094a79ec1656abaf762dc
Partial-Bug: #1739318