Remove the inherits parameter for the Resource object
In the past, the '/os-volumes_boot' API is inherited from the '/servers'
API, then the Resource object have the ability to get method from the
inherits controller. But now, we use the plain routes, all the methods
of '/os-volumes_boot' are listed in the route list now.

Partial implement bp api-extensions-merge-queens

Change-Id: Iff44b7756fd9a0f4aa78fa4306e66344011d0c2f